Remove oh file prefix, nullability of ids, rename status fields
