+++ backend/entities/office_hours/oh_section_entity.py
@@ -3,8 +3,8 @@
 from sqlalchemy import ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from backend.models.office_hours.oh_section import OfficeHoursSection
-from backend.models.office_hours.oh_section_details import OfficeHoursSectionDetails
+from backend.models.office_hours.section import OfficeHoursSection
+from backend.models.office_hours.section_details import OfficeHoursSectionDetails
 
 
 from ..entity_base import EntityBase
@@ -47,10 +47,7 @@ class OfficeHoursSectionEntity(EntityBase):
         Returns:
             OfficeHoursSectionEntity: Entity created from model
         """
-        return cls(
-            id=model.id,
-            title=model.title
-        )
+        return cls(id=model.id, title=model.title)
 
     def to_model(self) -> OfficeHoursSection:
         """
@@ -59,11 +56,8 @@ class OfficeHoursSectionEntity(EntityBase):
         Returns:
             OfficeHoursSection: `OfficeHoursSection` object from the entity
         """
-        return OfficeHoursSection(
-            id=self.id,
-            title=self.title
-        )
-    
+        return OfficeHoursSection(id=self.id, title=self.title)
+
     def to_details_model(self) -> OfficeHoursSectionDetails:
         """
         Converts a `OfficeHoursSectionEntity` object into a `OfficeHoursSectionDetails` model object
@@ -75,5 +69,5 @@ class OfficeHoursSectionEntity(EntityBase):
             id=self.id,
             title=self.title,
             sections=[section.to_model() for section in self.sections],
-            events= [event.to_model() for event in self.events]
+            events=[event.to_model() for event in self.events],
         )
