+++ backend/models/office_hours/event_status.py
@@ -3,7 +3,7 @@
 from pydantic import BaseModel
 from typing import Sequence
 
-from oh_ticket import OfficeHoursTicket
+from .ticket import OfficeHoursTicket
 
 __authors__ = ["Sadie Amato, Bailey DeSouza, Meghan Sun, Maddy Andrews"]
 __copyright__ = "Copyright 2024"
@@ -13,14 +13,14 @@ __license__ = "MIT"
 class TAOfficeHoursEventStatus(BaseModel):
     """Pydantic model to represent the status of a given event, including tickets in the queue, for a given TA."""
 
-    active_tickets: Sequence[OfficeHoursTicket]
-    num_open: int
-    num_pending: int
+    open_and_queued_tickets: Sequence[OfficeHoursTicket]
+    open_tickets_count: int
+    queued_tickets_count: int
 
 
 class StudentOfficeHoursEventStatus(BaseModel):
     """Pydantic model to represent the status of a given event, including their position in the queues, for a given student."""
 
-    num_open: int
-    num_pending: int
-    position: int
+    open_tickets_count: int
+    queued_tickets_count: int
+    queue_position: int
